+++ src/hooks/useProfilePicture/useProfilePicture.js
@@ -5,14 +5,15 @@ import {
   useOkapiKy,
 } from '@folio/stripes/core';
 
-import { PROFILE_PIC_API } from '../../../../../constants';
-import { isAValidUUID } from '../../../../util/util';
+import { PROFILE_PIC_API } from '../../constants';
+import { isAValidUUID } from '../../components/util/util';
 
 const useProfilePicture = ({ profilePictureId }, options = {}) => {
   const ky = useOkapiKy();
   const [namespace] = useNamespace({ key: 'get-profile-picture-of-a-user' });
   const {
     isFetching,
+    isLoading,
     data = {},
   } = useQuery(
     [namespace, profilePictureId],
@@ -26,6 +27,7 @@ const useProfilePicture = ({ profilePictureId }, options = {}) => {
   );
 
   return ({
+    isLoading,
     isFetching,
     profilePictureData: data?.profile_picture_blob,
   });